cmd/psm: Update the leader election configurations
Update the package-server-manager controller and update the leader
election intervals. This is in order to comply with the need for OCP
components to be able to withstand 60s of API server disruption on
SNO-enabled clusters.

For more information, see the following resources:
- https://github.com/openshift/library-go/pull/1104#discussion_r649313822
- https://bugzilla.redhat.com/show_bug.cgi?id=1985697

Alternative implementations include disabling leader election entirely
for SNO-enabled clusters. This implementation is centered around
dynamically querying for the Infrastructure/cluster singleton resource,
checking the HA/non-HA expectations being exposed, and setting leader
election properly. This implementation would still need to be careful
about how to handle transient errors and provide an escape hatch (e.g.
prefer enablement of leader election through a CLI flag, vs. the dynamic
value) that users can pass to the PSM deployment for failed upgrades.
