+++ cmd/package-server-manager/main.go
@@ -6,13 +6,9 @@ import (
 
 	"github.com/spf13/cobra"
 
-	configv1 "github.com/openshift/api/config/v1"
 	olmv1alpha1 "github.com/operator-framework/api/pkg/operators/v1alpha1"
 
 	"k8s.io/apimachinery/pkg/fields"
-	"k8s.io/apimachinery/pkg/runtime"
-	utilruntime "k8s.io/apimachinery/pkg/util/runtime"
-	clientgoscheme "k8s.io/client-go/kubernetes/scheme"
 	_ "k8s.io/client-go/plugin/pkg/client/auth"
 
 	ctrl "sigs.k8s.io/controller-runtime"
@@ -71,6 +67,9 @@ func run(cmd *cobra.Command, args []string) error {
 		LeaderElection:          !disableLeaderElection,
 		LeaderElectionNamespace: namespace,
 		LeaderElectionID:        leaderElectionConfigmapName,
+		RetryPeriod:             timeDurationPtr(defaultRetryPeriod),
+		RenewDeadline:           timeDurationPtr(defaultRenewDeadline),
+		LeaseDuration:           timeDurationPtr(defaultLeaseDuration),
 		HealthProbeBindAddress:  healthCheckAddr,
 		NewCache: cache.BuilderWithOptions(cache.Options{
 			SelectorsByObject: cache.SelectorsByObject{
@@ -113,13 +112,4 @@ func run(cmd *cobra.Command, args []string) error {
 	}
 
 	return nil
-}
-
-func setupScheme() *runtime.Scheme {
-	scheme := runtime.NewScheme()
-	utilruntime.Must(clientgoscheme.AddToScheme(scheme))
-	utilruntime.Must(configv1.Install(scheme))
-	utilruntime.Must(olmv1alpha1.AddToScheme(scheme))
-
-	return scheme
 }
